winnt.h: fix ARM64 CONTEXT Bvr/Wvr register types
The fallback CONTEXT struct incorrectly defined Bvr (Breakpoint Value
Registers) and Wvr (Watchpoint Value Registers) as DWORD (32-bit) instead
of DWORD64 (64-bit).

On ARM64:
- BCR/WCR (Control Registers) are 32-bit ✓
- BVR/WVR (Value Registers) are 64-bit ✓

This mismatch caused struct size and layout errors, potentially corrupting
debug register state when used with Windows debugging APIs.

+++ win32/include/winapi/winnt.h
@@ -2117,9 +2117,9 @@ typedef DWORD LCID;
       DWORD Fpcr;
       DWORD Fpsr;
       DWORD Bcr[8];
-      DWORD Bvr[8];
+      DWORD64 Bvr[8];
       DWORD Wcr[2];
-      DWORD Wvr[2];
+      DWORD64 Wvr[2];
     } CONTEXT,*PCONTEXT;
 #endif
